Show permissions dialog only for untrusted origins
