+++ src/main/permissionRequestHandler.js
@@ -1,63 +1,5 @@
 const {ipcMain} = require('electron');
 const {URL} = require('url');
-const fs = require('fs');
-
-const PERMISSION_GRANTED = 'granted';
-const PERMISSION_DENIED = 'denied';
-
-class PermissionManager {
-  constructor(file) {
-    this.file = file;
-    if (fs.existsSync(file)) {
-      try {
-        this.permissions = JSON.parse(fs.readFileSync(this.file, 'utf-8'));
-      } catch (err) {
-        console.error(err);
-        this.permissions = {};
-      }
-    } else {
-      this.permissions = {};
-    }
-  }
-
-  writeFileSync() {
-    fs.writeFileSync(this.file, JSON.stringify(this.permissions, null, '  '));
-  }
-
-  grant(origin, permission) {
-    if (!this.permissions[origin]) {
-      this.permissions[origin] = {};
-    }
-    this.permissions[origin][permission] = PERMISSION_GRANTED;
-    this.writeFileSync();
-  }
-
-  deny(origin, permission) {
-    if (!this.permissions[origin]) {
-      this.permissions[origin] = {};
-    }
-    this.permissions[origin][permission] = PERMISSION_DENIED;
-    this.writeFileSync();
-  }
-
-  clear(origin, permission) {
-    delete this.permissions[origin][permission];
-  }
-
-  isGranted(origin, permission) {
-    if (this.permissions[origin]) {
-      return this.permissions[origin][permission] === PERMISSION_GRANTED;
-    }
-    return false;
-  }
-
-  isDenied(origin, permission) {
-    if (this.permissions[origin]) {
-      return this.permissions[origin][permission] === PERMISSION_DENIED;
-    }
-    return false;
-  }
-}
 
 function dequeueRequests(requestQueue, permissionManager, origin, permission, status) {
   switch (status) {
@@ -88,8 +30,7 @@ function dequeueRequests(requestQueue, permissionManager, origin, permission, st
   }
 }
 
-function permissionRequestHandler(mainWindow, permissionFile) {
-  const permissionManager = new PermissionManager(permissionFile);
+function permissionRequestHandler(mainWindow, permissionManager) {
   const requestQueue = [];
   ipcMain.on('update-permission', (event, origin, permission, status) => {
     dequeueRequests(requestQueue, permissionManager, origin, permission, status);
@@ -113,7 +54,5 @@ function permissionRequestHandler(mainWindow, permissionFile) {
     mainWindow.webContents.send('request-permission', targetURL.origin, permission);
   };
 }
-
-permissionRequestHandler.PermissionManager = PermissionManager;
 
 module.exports = permissionRequestHandler;
